refactor time-utils location
