+++ src/constants.js
@@ -6,7 +6,6 @@ const STATE_OPEN = 'open'
 const STATE_CLOSED = 'closed'
 const STATE_CLOSED_NOT_PLANNED = 'not_planned'
 const ISSUES_EVENT_NAME = 'issues'
-const COMMITS_WITHOUT_PRS_KEY = -1
 
 module.exports = {
   ISSUE_LABEL,
@@ -15,5 +14,4 @@ module.exports = {
   STATE_CLOSED,
   STATE_CLOSED_NOT_PLANNED,
   ISSUES_EVENT_NAME,
-  COMMITS_WITHOUT_PRS_KEY,
 }
